LayoutTests/fast/css/legacy-opacity-styles.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../js/resources/js-test-pre.js"></script>
</head>
<body>
<div id="test-opacity" style="opacity: 0.2;"></div>
<div id="test-webkit-opacity" style="-webkit-opacity: 0.8;"></div>
<p id="description"></p>
<div id="console"></div>
<script>
description("This test checks that the legacy -apple-opacity, -khtml-opacity and -webkit-opacity styles are mapped to the opacity style.");

shouldBe("document.getElementById('test-opacity').style.getPropertyValue('opacity')", "'0.2'");
shouldBe("document.getElementById('test-opacity').style.getPropertyValue('-webkit-opacity')", "'0.2'");

shouldBe("document.getElementById('test-webkit-opacity').style.getPropertyValue('opacity')", "'0.8'");
shouldBe("document.getElementById('test-webkit-opacity').style.getPropertyValue('-webkit-opacity')", "'0.8'");
</script>
<script src="../js/resources/js-test-post.js"></script>
</body>
</html>
