third_party/WebKit/LayoutTests/fast/css/legacy-opacity-styles-expected.txt

This test checks that the legacy -webkit-opacity style is mapped to the opacity style.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.getElementById('test-opacity').style.getPropertyValue('opacity') is '0.2'
PASS document.getElementById('test-opacity').style.getPropertyValue('-webkit-opacity') is '0.2'
PASS document.getElementById('test-webkit-opacity').style.getPropertyValue('opacity') is '0.8'
PASS document.getElementById('test-webkit-opacity').style.getPropertyValue('-webkit-opacity') is '0.8'
PASS successfullyParsed is true

TEST COMPLETE
